+++ pyfar/dsp/filter/band_filter.py
